+++ tests/cli_tests.rs
@@ -74,7 +74,7 @@ fn test_logo_override_flag() {
     cmd.args(["--no-color", "--logo", "arch", "-m", "os"]);
     let assert = cmd.assert().success();
     let stdout = String::from_utf8(assert.get_output().stdout.clone()).unwrap();
-    assert!(stdout.contains("/\\"));
+    assert!(stdout.contains("+oooo"));
     assert!(stdout.contains("OS:"));
 }
 
@@ -168,18 +168,18 @@ fn test_installed_module_filtering() {
 #[test]
 fn test_various_logo_overrides() {
     let logos = [
-        ("debian", "_____"),
-        ("ubuntu", "---(_)"),
-        ("arch", "/\\"),
-        ("fedora", "_____"),
+        ("debian", "met$$$$$gg"),
+        ("ubuntu", "clooo"),
+        ("arch", "+oooo"),
+        ("fedora", "cccccccc"),
         ("mint", "___________"),
-        ("tux", ".--."),
-        ("windows11", "#######  #######"),
-        ("windows10", "+-------"),
+        ("tux", ".---."),
+        ("windows11", "################  ################"),
+        ("windows10", "..,:-==+??***"),
         ("windows7", "_.-;;-._"),
-        ("windows", "#######  #######"),
-        ("win11", "#######  #######"),
-        ("win10", "+-------"),
+        ("windows", "################  ################"),
+        ("win11", "################  ################"),
+        ("win10", "..,:-==+??***"),
         ("win7", "_.-;;-._"),
     ];